reserve/requirements.txt
certifi
chardet
cssselect
idna
lxml
requests
selenium
urllib3
pyyaml
